[android] Adapt test to upstream readobj output change.
This is a change that will be necessary for master-next. readobj seems
to report elf32-littlearm when compiling for Android ARMv7.

This is similar to commit f7cf5bde468e29ebe8239fbea53af39e7ecf4f5d which
is still available in master-next, but I think was (mistakenly) reverted
with #32265.