[Effects handling] Eliminate Context::getHandled().
We don't want to expose the notion of "handles everything", because
that's a honeypot for messing up 'async' checking when making a context
change for error-handling, and vice versa. Instead, provide an API to
produce a derived context where all errors are handled, and use it for
(e.g.) try/try!/try? and exhaustive do-catches.