[IRGen] Teach defineAlias to replace useds.
Previously, it was possible to alias an old LinkEntity which was marked
used (i.e. in LLVMUsed) or compiler used (i.e. in LLVMCompilerUsed).
The result of defining such an alias was a failure downstream when
writing out the lists of globals.

Here, the old value is removed from the list before it is invalidated.