[IRGen] Erase nonoverride descriptor on emission.
Previously, emitting the descriptor for a nonoverride method always
simply emitted, even if it had previously been emitted.  That was not a
problem before, but is now that class type context descriptors can be
reemitted upon encountering metadata prespecializations.

Here, the behavior is changed to delete the old definition before
emitting the new definition.