[FunctionBuilders] Implement graceful handling of pre-check failures
Function builder bodies are applied in several stages. First of them is `pre-check`
stage which is, among other things, responsible for member lookup and reference
resolution, if that action fails it would be diagnosed inline and affected
expression is going to be replaced with `ErrorExpr`.

This works fine in context of regular expressions but not function builders, because
matching of function builder bodies happens in the middle of type-checking of
enclosing context, so solver needs to implement graceful handling of such problems
and delay diagnostics until solver reaches a solution.

To accommodate that `PreCheckFunctionBuilderRequest`,
`ConstraintSystem::preCheckExpression`, and `TypeChecker::resolveDeclRefExpr`
have been adjusted to know explicitly whether diagnosing and AST mutation is permitted.
Solver is adjusted to ignore closure bodies which fail pre-check in their entirety.

Resolves: rdar://problem/65320500