[Concurrency] Add @asyncHandler attribute.
An asynchronous handler behaves externally like a synchronous function,
but internally it is handled like an asynchronous function, meaning
that it can execute operations with suspension points such as calling
other async functions.  The body of the function is executed just as if
it and its caller were async functions, except that the caller is
resumed when the callee reaches its first suspension point rather than
only when the callee exits.