Excise Overlay Loading From Clang Module Import Path
There is no reason for the clang importer to load Swift modules. It does
so today because the computation of its transitive import graph via
namelookup::getAllImports forces the load of corresponding Swift
modules.

In lieu of a primitive that we can use to load just the clang importer
graph, push resolution of the import graph into the callers.

This is less than ideal. We could probably use
clang::Module::getExportedModules here instead.