IRGen: avoid tripping assertions on a move constructor invocation
The moved instance has not yet emitted the calls, however, the instance
is not meant to actually emit any arguments for the call as it has been
moved.  This would break when the compiler actually emits a move
constructor invocation rather and fails to do a NVRO instead.

Patch originally by Patrick Deschênes!