[SIL] Add flag to SILFunctionType::Profile for async.
Previously, the flag was omitted, causing function types which were
otherwise the same to have the same id, leading to caching woes.

Here, the issue is fixed by adding the boolean flag to the id, ensuring
that types which differ only in that flag are still understood to be
different.